requirements.txt
Flask==1.0.2 

Flask-Cors==3.0.7

Flask-SQLAlchemy==2.4.1

python-dotenv==0.13.0

PyMySQL== 0.9.3